Rename app/lib/redis.rb to app/lib/redis_service.rb
Match filename to the constant it defines.
