+++ app/lib/redis_service.rb
